+++ src/models/character/search.rs
@@ -13,23 +13,23 @@ pub struct SearchResult {
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "PascalCase")]
 pub struct Pagination {
-  page_current: usize,
-  page_next: usize,
-  page_previous: usize,
-  page_total: usize,
-  total: usize,
+  page_current: u64,
+  page_next: u64,
+  page_previous: u64,
+  page_total: u64,
+  total: u64,
 }
 
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "PascalCase")]
 pub struct SearchCharacter {
   #[serde(rename = "ID")]
-  pub id: usize,
+  pub id: u64,
   pub name: String,
   pub server: World,
   #[serde(with = "url_serde")]
   pub avatar: Url,
   pub rank: Option<serde_json::Value>,
   pub rank_icon: Option<serde_json::Value>,
-  pub feasts: usize,
+  pub feasts: u64,
 }
